+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## v0.16-dev
 
+* JavaScript client backwards incompatible changes
+  * Socket params are now passed to `socket.connect()` instead of an option on the constructor.
+  * Socket params are now longer merged as default params for channel params. Use `connect/2` on the server to wire up default channel assigns.
+
+
 ## v0.15.0 (2015-7-27)
 
 See these [`0.14.x` to `0.15.0` upgrade instructions](https://gist.github.com/chrismccord/931373940f320bf41a50) to bring your existing apps up to speed.